feat(drivers): implement IO handling
The pinmapping for the Launchpad is represented with an enum(io_e).
Create a second enum(io_generic_e), define macros, and static register arrays
(e.g. port_dir_regs) to map the enum values to the corresponding registers
in a space-efficient way. Implement the individual functions to select functions,
set pupd resister,direction and output. The result is an abtraction that's much
nicer than accessing the crude register variables from TI.