feat(assert): assert handler and led driver
Create an assert handler suitable for microcontroller. Let it trigger
a software breakpoint (if debugger is attached) and then enter an
endless while loop blinking the test led. The handler will later be
extended with console tracing and turning off the motors.

Also, implement a simple led driver to avoid having to interact with
the IO functoins directly, but dont use this driver in assert handler
to avoid it to being called recursively.

Finally, wrap the __delay_cycles_ms intrinsic in a new define macro
BUSY_WAIT_ms that takes milliseconds (instead of cycles) as
argument.