build: add two makefile rules for analyzing flash space used.
MSP430F5529 has 128KB of flash, so care must be taken when programming
it as to not run out of this small space. The GCC-toolchain
has several tools for seeing how much space the code occupies,
add these as rules to the Makefile.

While at it, remove "inline" from some of the IO functions because
they apparently occupy more space as inlined. May lead to more CPU
cycles, but insignificant, and prefer space over speed in this case.