feat(trace): Use external printf implementation for trace
The standard library printf implemenattion is not appropriate for a
microcontroller. A microcontroller has no notion of stdout and
standard printf implenetation has a big memory footprint. To get
around this, use a more stripped down external implementation,
mpland/printf, and include it as git submodule(already included).
Use this to implement a trace function that can format strings.