feat(ir_remote): driver to decode IR signal
The robot must be able to decode a start signal sent at each match
start. This signal is sent from an IR transmitter and is picked up
by the IR receiver on board. Implement a driver that decodes the signal by
measuring the time between the pulses on the GPIO connected to the IR
receiver. Besides the start signal, also use this functionality to control
the robot with a remote control, which is useful during development.