feat(drivers): line detector qre1113 driver
Add a driver for the line detector qre1113. These line detectors are
connected to the microcontroller with one line each that are sampled
by the ADC driver. As such this is mostly a thin code layer for
readability.

NOTE: CURRENTLY, THIS COMMIT IS NOT FULLY FUNCTIONAL. THIS COMMIT
DOESN'T PROVIDE THE EXPECTED ADC OUTPUT. I AM WORKING ON
RESOLVING THE ISSUE.....