feat(app): Convert line detector voltages to a line position
There are four line detectors on the actual robot, and the combination
of their readings maps to a distinct line position. Add a function that
translate the readings into the line position.

NOTE: CURRENTLY, THIS CODE IS NOT FULLY FUNCTONAL.THIS COMMIT DOESN'T
PROVIDE THE EXPECTED THE ADC OUPUT.I AM WORKING ON RESOLVING THE ISSUE....