feat(drivers): VL53L0X and fix io_set_interrupt_trigger bug
The robot has five range sensors, three in the front and one on either
side. These are complex sensors and requires a comprehensive driver. Add
the required driver on top of the  I2C driver.

Also, io_set_interrupt_trigger is wrong, it sets rising when it's
supposed to be falling, vice versa. Fix this.